[modify] Flask 프로젝트 모듈화
- app.py: Flask 앱 초기화 및 Blueprint 등록
- routes.py: /speech_to_text, /chatbot/faq 라우팅 분리
- speech_to_text.py: Google STT 처리 로직 분리
- chatbot_faq.py: FAQ 기반 응답 처리 + TTS 분기 처리
- text_to_speech.py: Google TTS 처리 함수 분리
- db.py: DB 연결 및 쿼리 함수 공통화
- config.py: 환경변수 및 오디오 경로 설정
